Code cleaning
2 utils libraries added.
The standalone and mcp server exploit on utils libraries for less updates in testing.
